Bundle JS files using browserify
