+++ data/storage.js
@@ -1,0 +1,10 @@
+// Create a storage object:
+var storage = {
+	allStreets: [],
+	searchResults: [],
+	streetDetails: [],
+	streetDetailsCopy: [],
+	images: []
+};
+
+module.exports = storage;
