+++ views/layout.pug
@@ -1,0 +1,24 @@
+doctype html
+html(lang="nl")
+
+	head
+		block head
+			meta(charset="utf-8")
+			meta(name="viewport" content="width=device-width")
+			title OBA App Server Side
+			link(rel="manifest" href="/manifest.json")
+			link(rel="stylesheet" type="text/css" href="/css/style.css")
+
+	body
+		script(src="/bundle.js" async="true" defer="true")
+		script.
+			if ('serviceWorker' in navigator) {
+				navigator.serviceWorker.register('/serviceworker.js')
+					.then(function(registration) {
+						console.log('Browser supports Service workers', registration)
+					}).catch(function(err) {
+						console.log('Failed to register service worker')
+					});
+			} else {
+				console.log('No support for service workers')
+			}
